# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.getContactsResponse:
    properties:
      contacts:
        items:
          $ref: '#/definitions/whatsapp.ContactInfo'
        type: array
    type: object
...
  response.Error:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
...
  whatsapp.ContactInfo:
    properties:
      name:
        type: string
      phone:
        type: string
      picture:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8900
info:
  contact: {}
  description: API to handle multiple WhatsApp instances
  title: ZapMeow API
  version: "1.0"
paths:
...
  /{instanceId}/contacts:
    get:
      consumes:
      - application/json
      description: Retrieves all contacts for the specified WhatsApp instance.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of contacts
          schema:
            $ref: '#/definitions/handler.getContactsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get WhatsApp Contacts
      tags:
      - WhatsApp Contacts
...
swagger: "2.0"